+++ umbral-pre/src/pre.rs
@@ -150,7 +150,8 @@ pub fn reencrypt_with_rng(
     capsule: &Capsule,
     verified_kfrag: &VerifiedKeyFrag,
 ) -> VerifiedCapsuleFrag {
-    VerifiedCapsuleFrag::reencrypted(rng, capsule, &verified_kfrag.kfrag)
+    let kfrag = verified_kfrag.to_unverified();
+    VerifiedCapsuleFrag::reencrypted(rng, capsule, &kfrag)
 }
 
 /// A synonym for [`reencrypt_with_rng`] with the default RNG.
@@ -179,8 +180,7 @@ pub fn decrypt_reencrypted(
 ) -> Result<Box<[u8]>, ReencryptionError> {
     let cfrags: Vec<_> = verified_cfrags
         .iter()
-        .cloned()
-        .map(|vcfrag| vcfrag.cfrag)
+        .map(|vcfrag| vcfrag.to_unverified())
         .collect();
     let key_seed = capsule
         .open_reencrypted(receiving_sk, delegating_pk, &cfrags)